Make readByte return byte instead of int

+++ src/main/java/com/github/tth05/scnet/util/ByteBufferInputStream.java
@@ -21,7 +21,7 @@ public class ByteBufferInputStream {
         this.buf = buffer;
     }
 
-    public int readByte() {
+    public byte readByte() {
         return this.buf.get();
     }
